# docs/openapi.yml
{
  "openapi": "3.0.2",
  "info": {
    "title": "Observer API",
    "description": "Observer API server",
    "version": "0.1.0"
  },
  "paths": {
    "/account/me": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Me",
        "operationId": "get_me_account_me_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/account/confirm/{code}": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Confirm Account",
        "operationId": "confirm_account_account_confirm__code__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            },
            "name": "code",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/account/confirmation/resend": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Resend Confirmation",
        "operationId": "resend_confirmation_account_confirmation_resend_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/invites": {
      "get": {
        "tags": [
          "admin",
          "invites"
        ],
        "summary": "Get Invites",
        "operationId": "get_invites_admin_invites_get",
        "parameters": [
          {
            "description": "How many items to return?",
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "description": "How many items to return?",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          },
          {
            "description": "How many items to skip?",
            "required": false,
            "schema": {
              "title": "Offset",
              "type": "integer",
              "description": "How many items to skip?",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInvitesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin",
          "invites"
        ],
        "summary": "Create Invite",
        "operationId": "create_invite_admin_invites_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "X-Forwarded-For",
              "type": "string",
              "default": ""
            },
            "name": "X-Forwarded-For",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInviteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/invites/{code}": {
      "delete": {
        "tags": [
          "admin",
          "invites"
        ],
        "summary": "Delete Invite",
        "operationId": "delete_invite_admin_invites__code__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Code",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "code",
            "in": "path"
          },
          {
            "description": "If set then related user will be deleted",
            "required": false,
            "schema": {
              "title": "Delete User",
              "type": "boolean",
              "description": "If set then related user will be deleted",
              "default": false
            },
            "name": "delete_user",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Forwarded-For",
              "type": "string",
              "default": ""
            },
            "name": "X-Forwarded-For",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "title": "APIError",
        "required": [
          "code",
          "status_code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "data": {
            "title": "Data"
          }
        }
      },
...
      "ErrorCode": {
        "title": "ErrorCode",
        "enum": [
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict_error",
          "totp_error",
          "totp_invalid_backup_code_error",
          "totp_required_error",
          "totp_exists_error",
          "registration_error",
          "registrations_closed_error",
          "password_reset_code_expired_error",
          "weak_password_error",
          "invalid_password_error",
          "similar_passwords_error",
          "document_is_too_large_error",
          "unsupported_document_format",
          "confirmation_code_expired_error",
          "invite_expired_error",
          "internal_error",
          "bad_request"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "InvitePermissionRequest": {
        "title": "InvitePermissionRequest",
        "required": [
          "project_id"
        ],
        "type": "object",
        "properties": {
          "can_create": {
            "title": "Can Create",
            "type": "boolean",
            "description": "User can create records",
            "default": false
          },
          "can_read": {
            "title": "Can Read",
            "type": "boolean",
            "description": "User can create records",
            "default": true
          },
          "can_update": {
            "title": "Can Update",
            "type": "boolean",
            "description": "User can update records",
            "default": false
          },
          "can_delete": {
            "title": "Can Delete",
            "type": "boolean",
            "description": "User can delete records",
            "default": false
          },
          "can_create_projects": {
            "title": "Can Create Projects",
            "type": "boolean",
            "description": "User can create projects",
            "default": false
          },
          "can_read_documents": {
            "title": "Can Read Documents",
            "type": "boolean",
            "description": "User can read documents",
            "default": false
          },
          "can_read_personal_info": {
            "title": "Can Read Personal Info",
            "type": "boolean",
            "description": "User can read personal info",
            "default": false
          },
          "can_invite_members": {
            "title": "Can Invite Members",
            "type": "boolean",
            "description": "User can invite new members to projects",
            "default": false
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "User can create records"
          }
        }
      },
...
      "Role": {
        "title": "Role",
        "enum": [
          "admin",
          "staff",
          "consultant",
          "guest"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "UserInviteRequest": {
        "title": "UserInviteRequest",
        "required": [
          "email",
          "role",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string",
            "description": "User email to send a new invite",
            "format": "email"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ],
            "description": "Role of a user"
          },
          "permissions": {
            "title": "Permissions",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitePermissionRequest"
            },
            "description": "List of permissions to different projects"
          }
        }
      },
      "UserInviteResponse": {
        "title": "UserInviteResponse",
        "required": [
          "code",
          "user_id",
          "expires_at"
        ],
        "type": "object",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string",
            "description": "Invite code"
          },
          "user_id": {
            "title": "User Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "User ID"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string",
            "description": "Expiration datetime",
            "format": "date-time"
          }
        }
      },
      "UserInvitesResponse": {
        "title": "UserInvitesResponse",
        "required": [
          "total",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer",
            "description": "Total amount of invites"
          },
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserInviteResponse"
            },
            "description": "List of invites DESC ordered by expiration date"
          }
        }
      },
      "UserResponse": {
        "title": "UserResponse",
        "required": [
          "email",
          "role",
          "id"
        ],
        "type": "object",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string",
            "description": "E-mail address of a user",
            "format": "email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string",
            "description": "Full name of a user"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ],
            "description": "Role of a user"
          },
          "office_id": {
            "title": "Office Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Office ID to which user belongs"
          },
          "id": {
            "title": "Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "ID of user"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean",
            "description": "Is user active?",
            "default": true
          },
          "is_confirmed": {
            "title": "Is Confirmed",
            "type": "boolean",
            "description": "Is user confirmed?",
            "default": false
          },
          "mfa_enabled": {
            "title": "Mfa Enabled",
            "type": "boolean",
            "description": "Is MFA enabled for user?",
            "default": false
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}